+++ azure-storage-file/azure/storage/file/__init__.py
@@ -25,4 +25,5 @@ from .models import (
     CopyProperties,
     SharePermissions,
     FilePermissions,
+    DeleteSnapshot,
 )